"Fixes" character encoding bug
This "fixes" the character encoding bug introduced in the migration to a Gradle plugin, by replacing the special characters with plus and minus signs. This is not a great fix because it changes the UI, but this means when updating the plugin in the future we would not run into this issue again.